refactor: 디렉토리 구조 변경 및 네이밍 수정
