+++ src/main/java/domain/lotto/LottoConstant.java
@@ -1,6 +1,6 @@
 package domain.lotto;
 
-public class LottoConstantData {
+public class LottoConstant {
     public static final Integer MIN_LOTTO_NUMBER = 1;
     public static final Integer MAX_LOTTO_NUMBER = 45;
     public static final Integer LOTTO_NUMBER_SIZE = 6;
