+++ src/main/java/domain/lotto/WinningNumbers.java
@@ -26,14 +26,14 @@ public class WinningNumbers {
 
     private void validateNumberDuplication(List<Integer> lottoNumbers, Integer bonusNumber) {
         HashSet<Integer> hs = new HashSet<>(lottoNumbers);
-        if (hs.size() != LottoConstantData.LOTTO_NUMBER_SIZE) throw new IllegalArgumentException();
+        if (hs.size() != LottoConstant.LOTTO_NUMBER_SIZE) throw new IllegalArgumentException();
         if (hs.contains(bonusNumber)) throw new IllegalArgumentException();
     }
 
     private void validateNumberRange(List<Integer> lottoNumbers) {
         lottoNumbers
                 .forEach((number) -> {
-                    if (number < LottoConstantData.MIN_LOTTO_NUMBER || number > LottoConstantData.MAX_LOTTO_NUMBER)
+                    if (number < LottoConstant.MIN_LOTTO_NUMBER || number > LottoConstant.MAX_LOTTO_NUMBER)
                         throw new IllegalArgumentException();
                 });
     }
